% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwdual.R
\name{mixKWDual}
\alias{mixKWDual}
\title{Add title here.}
\usage{
mixKWDual(L, ...)
}
\arguments{
\item{L}{Matrix specifying the optimization problem to be solved.}

\item{...}{Additional optimization parameters passed to MOSEK. See
\code{\link[REBayes]{KWDual}} for details.}
}
\value{
\code{mixKWDual} returns a list with two components:

\item{x}{The solution to the optimization problems provided by 
  MOSEK.}

\item{status}{The MOSEK return status.}
}
\description{
Add more detailed description here. Explain that this
  solves the dual problem formulation using the MOSEK interior-point
  solver (see text from manuscript).
}
\examples{
# Add example here.

}
