% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixobjective.R
\name{mixobjective}
\alias{mixobjective}
\title{Compute the value of the objective at x; arguments L and w specify
the objective, and e is an additional constant that can be set to a
small, positive number (zero by default) to better ensure numerical
stability of the optimization.}
\usage{
mixobjective(L, w = rep(1, nrow(L)), x = rep(1, ncol(L)), e = 0)
}
\description{
Compute the value of the objective at x; arguments L and w specify
the objective, and e is an additional constant that can be set to a
small, positive number (zero by default) to better ensure numerical
stability of the optimization.
}
