\name{mixSQP-package}
\alias{mixSQP-package}
\alias{mixSQP}
\docType{package}
\title{
\packageTitle{mixSQP}
}
\description{
\packageDescription{mixSQP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mixSQP}
\packageIndices{mixSQP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mixSQP}

Maintainer: \packageMaintainer{mixSQP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
