% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{testdata}
\alias{testdata}
\title{testdata}
\usage{
testdata(n, m, mix_type = "mix_n")
}
\arguments{
\item{n}{nrow}

\item{m}{ncol}

\item{mix_type}{a type of mixture: "mix_n" for normal mixture, "mix_nt" for normal and t mixture.}
}
\description{
samples L by means of nonparametric empirical Bayes approach for estimation of mixture proportions.
}
\examples{
n  <- 1e5
m  <- 20
L  <- testdata(n,m, mix_type = "mix_n") # Create some simulated data

}
