% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebayes.R
\name{mixkwdual}
\alias{mixkwdual}
\title{Maximum likelihood estimation of mixture proportions}
\usage{
mixkwdual(L, w = rep(1, nrow(L)), ...)
}
\arguments{
\item{L}{a matrix of likelihoods (see \code{mixopt} for details).}

\item{w}{a vector of weights (see \code{mixopt} for details).}

\item{...}{Additional arguments passed to \code{\link[REBayes]{KWDual}}.}
}
\description{
Maximum likelihood estimation of mixture proportions
}
\details{
This function is simply a convenient wrapper to REBayes::KWDual to solve the same
problem as \code{mixopt}. See the \code{mixopt} documentation for details.
}
