% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{mixobjective}
\alias{mixobjective}
\title{Compute objective function optimized by mixsqp}
\usage{
mixobjective(L, x, w = rep(1, nrow(L)))
}
\arguments{
\item{L}{Matrix specifying the optimization problem to be solved.
In the context of mixture-model fitting, \code{L[j,k]} should be
the value of the kth mixture component density at the jth data
point. \code{L} should be a numeric matrix with at least two
columns, with all entries positive and finite (and not
missing). For large matrices, it is preferrable that the matrix is
stored in double-precision; see \code{\link{storage.mode}}.}

\item{x}{The point at which the objective is evaluated in
\code{mixobjective}; see argument \code{x0} to \code{mixsqp} for details.}

\item{w}{An optional numeric vector, with one entry for each row of
\code{L}, specifying the "weights" associated with the rows of
\code{L}. All weights must be finite, non-negative and not
missing. Internally the weights are normalized to sum to one, which does not
change the problem, but does change the value of the objective function reported.  
By default, all weights are equal.}
}
\description{
Compute objective function optimized by mixsqp
}
\details{
see \code{mixsqp} for details of the objective function
}
