% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixsqp-package.R
\docType{package}
\name{mixsqp-package}
\alias{mixsqp-package}
\title{mixsqp: Sequential Quadratic Programming for Fast Maximum-Likelihood
Estimation of Mixture Proportions}
\description{
Provides optimization algorithms based on sequential quadratic
programming (SQP) for maximum likelihood estimation of the mixture
proportions in a finite mixture model where the component densities
are known. To learn more, visit
\url{https://github.com/stephenslab/mixsqp}, and see the help for
function \code{\link{mixsqp}}.
}
