% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixsqp.R
\name{mixSQP}
\alias{mixSQP}
\title{mixSQP}
\usage{
mixSQP(L, w = rep(1, nrow(L)), x0 = rep(1, ncol(L)), convtol = 1e-08,
  sparsetol = 1e-08, eps = .Machine$double.eps, maxitersqp = 1000,
  maxiteractiveset = 100, verbose = TRUE)
}
\arguments{
\item{L}{Matrix specifying the optimization problem to be
solved. It should be a numeric matrix with positive entries, and,
ideally, double-precision.}

\item{w}{A numeric vector, with one entry for each row of \code{L},
specifying the "weights" associated with the rows of \code{L}. All
weights must be positive. It is assumed the weights sum to 1; if
not, they will automatically be normalized to sum to 1. By default,
all rows of \code{L} are assigned the same weight.}

\item{x0}{An initial estimate of the solution to the optimization
problem. It should contain only non-negative values, and the
entries should sum to 1; if not, the vector is automatically
normalized to sum to 1. By default, \code{x0} is the vector with
all equal entries.}

\item{convtol}{A convergence tolerance used for algorithm's
convergence criterion.}

\item{sparsetol}{A tolerance used for determining active indices.}

\item{eps}{A small constant to safeguard from a numerical issue
(default 1e-6).}

\item{maxitersqp}{Maximum number of SQP iterations; that is, the
maximum number of quadratic subproblems that will be solved by the
active-set method.}

\item{maxiteractiveset}{Maximum number of active-set iterations
taken to solve each of the quadratic subproblems.}

\item{verbose}{If \code{verbose = TRUE}, print progress of algorithm
to console.}
}
\value{
Returns a solution x (in the current version).
}
\description{
mixSQP solves a convex optimization problem in
  https://arxiv.org/abs/1806.01412.
  It implements a sequential quadratic programming with
  active-set subproblem solvers. For gigantic data, use low-rank
  approximation to speed up the computation.

When L is a (n) by (m) matrix of nonnegative entries, mixSQP
maximizes the following objective function \deqn{f(x) = \sum_j w_j
log (\sum_k L_jk x_k)} subject to the (unit) probability simplex
constraint \deqn{\sum_k x_k = 1, x_k \ge 0} Without loss of
generality \eqn{\sum_j w_j = 1} is required.
}
\examples{
n  <- 1e5
m  <- 20
w  <- rep(1,n)/n; # weights
L  <- testdata(n,m, mix_type = "mix_n") # Create some simulated data
x0 <- rep(1,m)/m # initialization
out <- mixSQP(L,x0,w)

}
